Add GitHub Actions workflow for ASP.NET app deployment to Azure
This commit introduces a new GitHub Actions workflow named "Build and deploy ASP app to Azure Web App - webapp4682". The workflow, triggered on push events to the master branch or manually, consists of two jobs: `build` and `deploy`. The `build` job checks out the code, sets up MSBuild and NuGet, restores NuGet packages, builds the application, and uploads the published files as an artifact. The `deploy` job, dependent on `build`, runs on the latest Windows version, downloads the artifact, logs into Azure, and deploys the application to 'webapp4682' Azure Web App in the 'Production' slot. The environment URL for the `deploy` job is set to the output URL of the Azure Web App deployment step. Write permissions for id-tokens have been granted to the `deploy` job for JWT requests.